add Golf, Heartrate, Hydration, and activities
