+++ src/garmin/common/HydrationUtils.ts
@@ -3,3 +3,9 @@ export function convertMLToOunces(valueInML: number): number {
     const valueInOunces = valueInML * conversionFactor;
     return valueInOunces;
 }
+
+export function convertOuncesToML(ounces: number): number {
+    const ouncesToMillilitersConversionFactor = 29.5735;
+    const milliliters = ounces * ouncesToMillilitersConversionFactor;
+    return milliliters;
+}
